refactor: code clean up and component breakdown
